refactor: move AST indexing into core engine
