+++ apps/api/Cargo.toml
@@ -31,10 +31,3 @@ dotenvy = "0.15"
 dirs = "5.0"
 clap = { workspace = true }
 sha2 = "0.10"
-tree-sitter = "0.26"
-tree-sitter-rust = "0.24.2"
-tree-sitter-javascript = "0.25"
-tree-sitter-typescript = "0.23.2"
-tree-sitter-python = "0.25"
-tree-sitter-go = "0.25"
-tree-sitter-java = "0.23.5"
