+++ apps/api/src/api/system/handlers.rs
@@ -5,21 +5,13 @@ use axum::{
     response::{IntoResponse, Response},
     Json,
 };
+use core_engine::build_code_ast_artifacts;
 use infra::utils::debug_logging::DebugLogger;
 use serde::Deserialize;
 use serde_json::{json, Value};
-use sha2::{Digest, Sha256};
-use std::{
-    collections::{BTreeMap, BTreeSet, HashSet},
-    fs::{self, File},
-    io::Write,
-    path::{Path, PathBuf},
-    process::Command,
-    time::{SystemTime, UNIX_EPOCH},
-};
+use std::{collections::HashSet, path::PathBuf};
 use tokio_util::io::ReaderStream;
 use tracing::{info, warn};
-use tree_sitter::{Language, Node, Parser};
 
 use crate::api::dto::ApiResponse;
 use crate::error::ApiError;
@@ -27,6 +19,8 @@ use crate::{app_state::AppState, error::ApiResult};
 
 use super::diagnostics;
 use super::skills;
+
+const PROJECT_WIKI_AST_DIR: &str = ".atmos/wiki/_ast";
 
 #[derive(Deserialize)]
 pub struct KillTmuxSessionPayload {
@@ -187,23 +181,22 @@ pub async fn kill_project_wiki_window(
 
 /// POST /api/system/project-wiki-ast/:workspace_id
 pub async fn build_project_wiki_ast(
-    State(_state): State<AppState>,
-    axum::extract::Path(_workspace_id): axum::extract::Path<String>,
+    State(state): State<AppState>,
+    axum::extract::Path(workspace_id): axum::extract::Path<String>,
     Json(payload): Json<BuildProjectWikiAstPayload>,
 ) -> ApiResult<Json<ApiResponse<Value>>> {
-    let project_path = PathBuf::from(payload.project_path.trim());
-    if !project_path.is_dir() {
-        return Err(ApiError::BadRequest(format!(
-            "Invalid project path: {}",
-            project_path.display()
-        )));
-    }
+    let project_path =
+        resolve_project_wiki_ast_root(&state, &workspace_id, PathBuf::from(payload.project_path))
+            .await?;
 
     let project_path_for_worker = project_path.clone();
-    let result = tokio::task::spawn_blocking(move || build_ast_artifacts(&project_path_for_worker))
-        .await
-        .map_err(|e| ApiError::InternalError(format!("AST worker join failed: {}", e)))?
-        .map_err(|e| ApiError::InternalError(e.to_string()))?;
+    let ast_dir_for_worker = project_path.join(PROJECT_WIKI_AST_DIR);
+    let result = tokio::task::spawn_blocking(move || {
+        build_code_ast_artifacts(&project_path_for_worker, &ast_dir_for_worker)
+    })
+    .await
+    .map_err(|e| ApiError::InternalError(format!("AST worker join failed: {}", e)))?
+    .map_err(|e| ApiError::InternalError(e.to_string()))?;
 
     let logger = DebugLogger::new("project_wiki");
     logger.log(
@@ -211,7 +204,9 @@ pub async fn build_project_wiki_ast(
         "ast_index_built",
         Some(json!({
             "projectPath": project_path.display().to_string(),
+            "discoveredFiles": result.discovered_files,
             "indexedFiles": result.indexed_files,
+            "skippedFiles": result.skipped_files,
             "symbolCount": result.symbol_count,
             "relationCount": result.relation_count,
             "commitHash": result.commit_hash,
@@ -220,514 +215,91 @@ pub async fn build_project_wiki_ast(
 
     Ok(Json(ApiResponse::success(json!({
         "success": true,
+        "discovered_files": result.discovered_files,
         "indexed_files": result.indexed_files,
+        "skipped_files": result.skipped_files,
         "symbol_count": result.symbol_count,
         "relation_count": result.relation_count,
-        "ast_dir": ".atmos/wiki/_ast",
+        "ast_dir": PROJECT_WIKI_AST_DIR,
         "commit_hash": result.commit_hash,
     }))))
 }
 
-#[derive(Debug)]
-struct AstBuildResult {
-    indexed_files: usize,
-    symbol_count: usize,
-    relation_count: usize,
-    commit_hash: Option<String>,
-}
+async fn resolve_project_wiki_ast_root(
+    state: &AppState,
+    context_id: &str,
+    requested_path: PathBuf,
+) -> ApiResult<PathBuf> {
+    let requested_path = canonical_project_dir(requested_path)?;
+    let mut allowed_roots = Vec::new();
 
-fn build_ast_artifacts(project_root: &Path) -> Result<AstBuildResult, std::io::Error> {
-    let wiki_ast_dir = project_root.join(".atmos").join("wiki").join("_ast");
-    let shard_dir = wiki_ast_dir.join("files");
-    fs::create_dir_all(&wiki_ast_dir)?;
-    fs::create_dir_all(&shard_dir)?;
-
-    let mut symbols_file = File::create(wiki_ast_dir.join("symbols.jsonl"))?;
-    let mut relations_file = File::create(wiki_ast_dir.join("relations.jsonl"))?;
-
-    let mut source_files = Vec::new();
-    collect_source_files(project_root, &mut source_files)?;
-
-    let mut symbol_count = 0usize;
-    let mut relation_count = 0usize;
-    let mut file_index = Vec::new();
-
-    for file_path in &source_files {
-        let rel = file_path
-            .strip_prefix(project_root)
-            .unwrap_or(file_path)
-            .to_string_lossy()
-            .replace('\\', "/");
-
-        let content = match fs::read_to_string(file_path) {
-            Ok(c) => c,
-            Err(_) => continue,
-        };
-
-        let Some(lang_kind) = language_kind_for_file(file_path) else {
-            continue;
-        };
-
-        let parsed = parse_with_embedded_parser(&content, lang_kind);
-        let (symbols, relations, parser_warning) = match parsed {
-            Ok(result) => result,
-            Err(err) => (Vec::new(), Vec::new(), Some(err)),
-        };
-
-        for symbol in &symbols {
-            writeln!(symbols_file, "{}", serde_json::to_string(symbol)?)?;
-            symbol_count += 1;
-        }
-        for relation in &relations {
-            writeln!(relations_file, "{}", serde_json::to_string(relation)?)?;
-            relation_count += 1;
-        }
-
-        let shard_name = format!("{}.json", hash_path(&rel));
-        let shard_rel_path = format!("files/{}", shard_name);
-        let shard_full_path = shard_dir.join(shard_name);
-        let shard_payload = json!({
-            "file": rel,
-            "language": lang_kind.as_str(),
-            "symbol_count": symbols.len(),
-            "relation_count": relations.len(),
-            "symbols": symbols,
-            "relations": relations,
-            "parser_warning": parser_warning,
-        });
-        fs::write(
-            shard_full_path,
-            serde_json::to_string_pretty(&shard_payload)?,
-        )?;
-
-        file_index.push(json!({
-            "file": rel,
-            "language": lang_kind.as_str(),
-            "symbol_count": shard_payload["symbol_count"],
-            "relation_count": shard_payload["relation_count"],
-            "shard": shard_rel_path,
-            "parser_warning": parser_warning,
-        }));
-    }
-
-    let commit_hash = current_git_commit(project_root);
-    let now_unix = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_secs())
-        .unwrap_or_default();
-
-    let language_count = count_languages(&source_files);
-    let hierarchy = build_hierarchy_index(
-        &file_index
-            .iter()
-            .filter_map(|v| v.get("file").and_then(Value::as_str))
-            .map(ToOwned::to_owned)
-            .collect::<Vec<_>>(),
-    );
-    let status = json!({
-        "generated_at_unix": now_unix,
-        "commit_hash": commit_hash,
-        "indexed_files": source_files.len(),
-        "symbol_count": symbol_count,
-        "relation_count": relation_count,
-        "languages": language_count,
-        "index_files": {
-            "global_symbols": "symbols.jsonl",
-            "global_relations": "relations.jsonl",
-            "file_index": "index.json",
-            "hierarchy": "hierarchy.json"
-        },
-        "generator": "api.system.build_project_wiki_ast.tree_sitter_embedded"
-    });
-    fs::write(
-        wiki_ast_dir.join("_status.json"),
-        serde_json::to_string_pretty(&status)?,
-    )?;
-    fs::write(
-        wiki_ast_dir.join("index.json"),
-        serde_json::to_string_pretty(&json!({
-            "generated_at_unix": now_unix,
-            "commit_hash": commit_hash,
-            "files": file_index,
-            "progressive_disclosure": {
-                "entrypoint": "index.json",
-                "first_step": "Read hierarchy.json to select relevant dirs/files",
-                "second_step": "Open only needed file shards under _ast/files/*.json",
-                "avoid": "Do not load all shard files at once"
-            }
-        }))?,
-    )?;
-    fs::write(
-        wiki_ast_dir.join("hierarchy.json"),
-        serde_json::to_string_pretty(&hierarchy)?,
-    )?;
-
-    Ok(AstBuildResult {
-        indexed_files: source_files.len(),
-        symbol_count,
-        relation_count,
-        commit_hash,
-    })
-}
-
-fn collect_source_files(root: &Path, out: &mut Vec<PathBuf>) -> Result<(), std::io::Error> {
-    let mut stack = vec![root.to_path_buf()];
-    while let Some(dir) = stack.pop() {
-        for entry in fs::read_dir(&dir)? {
-            let entry = entry?;
-            let path = entry.path();
-            let file_name = path
-                .file_name()
-                .map(|s| s.to_string_lossy().to_string())
-                .unwrap_or_default();
-
-            if entry.file_type()?.is_dir() {
-                if should_skip_dir(&file_name) {
-                    continue;
-                }
-                stack.push(path);
-                continue;
-            }
-
-            if is_source_file(&path) {
-                out.push(path);
-            }
-        }
-    }
-    Ok(())
-}
-
-fn should_skip_dir(name: &str) -> bool {
-    matches!(
-        name,
-        ".git" | "node_modules" | "target" | "dist" | "build" | ".next" | ".turbo" | ".atmos"
-    )
-}
-
-fn is_source_file(path: &Path) -> bool {
-    matches!(
-        path.extension().and_then(|s| s.to_str()),
-        Some("rs" | "ts" | "tsx" | "js" | "jsx" | "py" | "go" | "java")
-    )
-}
-
-#[derive(Clone, Copy)]
-enum TsLanguageKind {
-    Rust,
-    JavaScript,
-    TypeScript,
-    Tsx,
-    Python,
-    Go,
-    Java,
-}
-
-impl TsLanguageKind {
-    fn as_str(self) -> &'static str {
-        match self {
-            TsLanguageKind::Rust => "rust",
-            TsLanguageKind::JavaScript => "javascript",
-            TsLanguageKind::TypeScript => "typescript",
-            TsLanguageKind::Tsx => "tsx",
-            TsLanguageKind::Python => "python",
-            TsLanguageKind::Go => "go",
-            TsLanguageKind::Java => "java",
-        }
-    }
-}
-
-fn language_kind_for_file(path: &Path) -> Option<TsLanguageKind> {
-    match path.extension().and_then(|s| s.to_str()) {
-        Some("rs") => Some(TsLanguageKind::Rust),
-        Some("js") | Some("jsx") => Some(TsLanguageKind::JavaScript),
-        Some("ts") => Some(TsLanguageKind::TypeScript),
-        Some("tsx") => Some(TsLanguageKind::Tsx),
-        Some("py") => Some(TsLanguageKind::Python),
-        Some("go") => Some(TsLanguageKind::Go),
-        Some("java") => Some(TsLanguageKind::Java),
-        _ => None,
-    }
-}
-
-fn language_for(kind: TsLanguageKind) -> Language {
-    match kind {
-        TsLanguageKind::Rust => tree_sitter_rust::LANGUAGE.into(),
-        TsLanguageKind::JavaScript => tree_sitter_javascript::LANGUAGE.into(),
-        TsLanguageKind::TypeScript => tree_sitter_typescript::LANGUAGE_TYPESCRIPT.into(),
-        TsLanguageKind::Tsx => tree_sitter_typescript::LANGUAGE_TSX.into(),
-        TsLanguageKind::Python => tree_sitter_python::LANGUAGE.into(),
-        TsLanguageKind::Go => tree_sitter_go::LANGUAGE.into(),
-        TsLanguageKind::Java => tree_sitter_java::LANGUAGE.into(),
-    }
-}
-
-fn parse_with_embedded_parser(
-    content: &str,
-    lang_kind: TsLanguageKind,
-) -> Result<(Vec<Value>, Vec<Value>, Option<String>), String> {
-    let mut parser = Parser::new();
-    parser
-        .set_language(&language_for(lang_kind))
-        .map_err(|e| format!("set_language failed: {}", e))?;
-
-    let tree = parser
-        .parse(content, None)
-        .ok_or_else(|| "parser returned no tree".to_string())?;
-
-    let mut symbols = Vec::new();
-    let mut relations = Vec::new();
-
-    let mut cursor = tree.walk();
-    let mut stack = vec![tree.root_node()];
-
-    while let Some(node) = stack.pop() {
-        if node.child_count() > 0 {
-            for child in node.children(&mut cursor) {
-                stack.push(child);
-            }
-        }
-
-        if !node.is_named() {
-            continue;
-        }
-
-        let kind = node.kind();
-        if is_symbol_node(kind) {
-            let name = node
-                .child_by_field_name("name")
-                .and_then(|n| node_text(content, n))
-                .unwrap_or_else(|| short_node_preview(content, node));
-            let point = node.start_position();
-            symbols.push(json!({
-                "kind": kind,
-                "name": name,
-                "line": point.row + 1,
-                "column": point.column + 1
-            }));
-        }
-
-        if is_relation_node(kind) {
-            let point = node.start_position();
-            relations.push(json!({
-                "kind": kind,
-                "line": point.row + 1,
-                "column": point.column + 1,
-                "snippet": short_node_preview(content, node)
-            }));
-        }
-    }
-
-    // Fallback if grammar did not capture enough meaningful nodes
-    if symbols.is_empty() && relations.is_empty() {
-        for (line_no, raw_line) in content.lines().enumerate() {
-            let line = raw_line.trim();
-            if let Some((kind, name)) = parse_symbol_line(line) {
-                symbols.push(json!({
-                    "kind": kind,
-                    "name": name,
-                    "line": line_no + 1,
-                    "column": 1
-                }));
-            }
-            if let Some(relation) = parse_relation_line(line) {
-                relations.push(json!({
-                    "kind": relation,
-                    "line": line_no + 1,
-                    "column": 1,
-                    "snippet": line.chars().take(140).collect::<String>()
-                }));
-            }
-        }
-    }
-
-    Ok((symbols, relations, None))
-}
-
-fn is_symbol_node(kind: &str) -> bool {
-    matches!(
-        kind,
-        "function_item"
-            | "function_declaration"
-            | "method_definition"
-            | "class_definition"
-            | "class_declaration"
-            | "struct_item"
-            | "enum_item"
-            | "trait_item"
-            | "interface_declaration"
-            | "type_alias_declaration"
-            | "impl_item"
-            | "method_declaration"
-            | "type_declaration"
-            | "function_definition"
-            | "class_specifier"
-            | "interface_type"
-    )
-}
-
-fn is_relation_node(kind: &str) -> bool {
-    matches!(
-        kind,
-        "use_declaration"
-            | "import_declaration"
-            | "import_statement"
-            | "import_from_statement"
-            | "mod_item"
-            | "extends_clause"
-            | "implements_clause"
-            | "call_expression"
-            | "scoped_identifier"
-    )
-}
-
-fn node_text(source: &str, node: Node) -> Option<String> {
-    source
-        .get(node.byte_range())
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn short_node_preview(source: &str, node: Node) -> String {
-    node_text(source, node)
-        .unwrap_or_default()
-        .chars()
-        .take(140)
-        .collect()
-}
-
-fn parse_symbol_line(line: &str) -> Option<(&'static str, String)> {
-    let symbol_prefixes = [
-        ("pub fn ", "function"),
-        ("fn ", "function"),
-        ("pub struct ", "struct"),
-        ("struct ", "struct"),
-        ("pub enum ", "enum"),
-        ("enum ", "enum"),
-        ("pub trait ", "trait"),
-        ("trait ", "trait"),
-        ("export function ", "function"),
-        ("function ", "function"),
-        ("export class ", "class"),
-        ("class ", "class"),
-        ("interface ", "interface"),
-        ("export interface ", "interface"),
-        ("type ", "type"),
-        ("export type ", "type"),
-    ];
-
-    for (prefix, kind) in symbol_prefixes {
-        if let Some(rest) = line.strip_prefix(prefix) {
-            let name = rest
-                .split(['(', '{', '<', ' ', ':', '='])
-                .next()
-                .unwrap_or_default()
-                .trim();
-            if !name.is_empty() {
-                return Some((kind, name.to_string()));
-            }
-        }
-    }
-    None
-}
-
-fn parse_relation_line(line: &str) -> Option<&'static str> {
-    if line.starts_with("use ")
-        || line.starts_with("pub use ")
-        || line.starts_with("import ")
-        || line.starts_with("from ")
+    if let Some(workspace) = state
+        .workspace_service
+        .get_workspace(context_id.to_string())
+        .await
+        .map_err(|e| ApiError::InternalError(e.to_string()))?
     {
-        return Some("import");
-    }
-    if line.starts_with("impl ") || line.contains(" implements ") {
-        return Some("implementation");
-    }
-    if line.starts_with("mod ") || line.starts_with("pub mod ") {
-        return Some("module");
-    }
-    if line.contains('(') && line.contains(')') && line.ends_with(';') {
-        return Some("call");
-    }
-    None
-}
-
-fn count_languages(files: &[PathBuf]) -> Vec<String> {
-    let mut set = BTreeSet::new();
-    for file in files {
-        if let Some(ext) = file.extension().and_then(|s| s.to_str()) {
-            set.insert(ext.to_string());
+        if !workspace.local_path.is_empty() {
+            allowed_roots.push(PathBuf::from(workspace.local_path));
+        }
+        if let Some(project) = state
+            .project_service
+            .get_project(workspace.model.project_guid)
+            .await
+            .map_err(|e| ApiError::InternalError(e.to_string()))?
+        {
+            allowed_roots.push(PathBuf::from(project.main_file_path));
         }
     }
-    set.into_iter().collect()
-}
 
-fn hash_path(path: &str) -> String {
-    let mut hasher = Sha256::new();
-    hasher.update(path.as_bytes());
-    format!("{:x}", hasher.finalize())
-}
-
-fn build_hierarchy_index(files: &[String]) -> Value {
-    #[derive(Default)]
-    struct DirNode {
-        files: Vec<String>,
-        children: BTreeMap<String, DirNode>,
+    if let Some(project) = state
+        .project_service
+        .get_project(context_id.to_string())
+        .await
+        .map_err(|e| ApiError::InternalError(e.to_string()))?
+    {
+        allowed_roots.push(PathBuf::from(project.main_file_path));
     }
 
-    fn insert_path(root: &mut DirNode, path: &str) {
-        let mut parts = path.split('/').collect::<Vec<_>>();
-        let file_name = parts.pop().unwrap_or_default().to_string();
-        let mut node = root;
-        for seg in parts {
-            node = node.children.entry(seg.to_string()).or_default();
+    if allowed_roots.is_empty() {
+        return Err(ApiError::BadRequest(format!(
+            "Unknown workspace or project: {}",
+            context_id
+        )));
+    }
+
+    for allowed_root in allowed_roots {
+        if let Ok(allowed_root) = canonical_project_dir(allowed_root) {
+            if requested_path == allowed_root {
+                return Ok(requested_path);
+            }
+            if requested_path.starts_with(&allowed_root) {
+                return Ok(allowed_root);
+            }
         }
-        node.files.push(file_name);
     }
 
-    fn to_json(path: &str, node: &DirNode) -> Value {
-        let children = node
-            .children
-            .iter()
-            .map(|(name, child)| {
-                let next_path = if path.is_empty() {
-                    name.to_string()
-                } else {
-                    format!("{}/{}", path, name)
-                };
-                to_json(&next_path, child)
-            })
-            .collect::<Vec<_>>();
-        json!({
-            "path": if path.is_empty() { "." } else { path },
-            "files": node.files,
-            "children": children
-        })
-    }
-
-    let mut root = DirNode::default();
-    for file in files {
-        insert_path(&mut root, file);
-    }
-    to_json("", &root)
+    Err(ApiError::BadRequest(
+        "Project path is outside the resolved workspace/project root".to_string(),
+    ))
 }
 
-fn current_git_commit(project_root: &Path) -> Option<String> {
-    let output = Command::new("git")
-        .arg("-C")
-        .arg(project_root)
-        .arg("rev-parse")
-        .arg("HEAD")
-        .output()
-        .ok()?;
-    if !output.status.success() {
-        return None;
+fn canonical_project_dir(path: PathBuf) -> ApiResult<PathBuf> {
+    let trimmed = path.to_string_lossy().trim().to_string();
+    let path = PathBuf::from(trimmed);
+    if !path.is_dir() {
+        return Err(ApiError::BadRequest(format!(
+            "Invalid project path: {}",
+            path.display()
+        )));
     }
-    let commit = String::from_utf8_lossy(&output.stdout).trim().to_string();
-    if commit.is_empty() {
-        None
-    } else {
-        Some(commit)
-    }
+    std::fs::canonicalize(&path).map_err(|e| {
+        ApiError::BadRequest(format!(
+            "Failed to resolve project path {}: {}",
+            path.display(),
+            e
+        ))
+    })
 }
 
 /// GET /api/system/code-review-window/:workspace_id
