+++ crates/core-engine/Cargo.toml
@@ -12,5 +12,13 @@ serde_json = "1.0"
 base64 = "0.22"
 regex = "1.12.3"
 ignore = "0.4"
+sha2 = "0.10"
+tree-sitter = "0.26"
+tree-sitter-rust = "0.24.2"
+tree-sitter-javascript = "0.25"
+tree-sitter-typescript = "0.23.2"
+tree-sitter-python = "0.25"
+tree-sitter-go = "0.25"
+tree-sitter-java = "0.23.5"
 serde.workspace = true
 reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
